ucu_chatbot_project/requirements.txt
Django>=4.2
google-genai>=0.6.0
python-dotenv
